ipq806x: clk: Fix compiler warnings
Remove compiler warning in clk and fab scaling.

+++ include/linux/fab_scaling.h
@@ -20,7 +20,7 @@
 
 struct fab_scaling_info {
 	struct clk *clk;
-	unsigned long idle_freq;
+	u32 idle_freq;
 };
 
 /**
